Fixed cache usage on fibonacciNthNumber Refactored fibonacciNthNumber to use BigInteger and accept higher N values

+++ src/main/kotlin/com/github/ricardoebbers/pricechecker/domain/service/LogicChallengesService.kt
@@ -1,9 +1,10 @@
 package com.github.ricardoebbers.pricechecker.domain.service
 
 import com.github.ricardoebbers.pricechecker.domain.enums.MoneyBill
+import java.math.BigInteger
 
 interface LogicChallengesService {
-    fun fibonacciNthNumber(n: Int): Long
+    fun fibonacciNthNumber(n: Int): BigInteger
     fun leastMoneyBillChange(changeValue: Int): Map<MoneyBill, Int>
     fun sumEvenNumbers(numbersSequence: List<Int>): Int
 }
